Create code_
This code defines three variables (diamondCount, rubyCount, and sapphireCount) to keep track of the number of each gem mined. Three functions (mineDiamond, mineRuby, and mineSapphire) are defined, each of which increments the count of the corresponding gem and logs a message to the console. Finally, the code calls the mining functions a few times to simulate mining, and then logs the final counts.